Separate vehicle logic from EntityManager
