+++ src/OpenLoco/src/Entities/EntityManager.cpp
@@ -8,7 +8,6 @@
 #include "Localisation/StringIds.h"
 #include "Map/Tile.h"
 #include "SceneManager.h"
-#include "Vehicles/Vehicle.h"
 #include <OpenLoco/Console/Console.h>
 #include <OpenLoco/Core/LocoFixedVector.hpp>
 #include <OpenLoco/Interop/Interop.hpp>
@@ -99,12 +98,6 @@ namespace OpenLoco::EntityManager
     uint16_t getListCount(const EntityListType list)
     {
         return rawListCounts()[static_cast<size_t>(list)];
-    }
-
-    template<>
-    Vehicles::VehicleHead* first()
-    {
-        return get<Vehicles::VehicleHead>(firstId(EntityListType::vehicleHead));
     }
 
     template<>
@@ -300,18 +293,6 @@ namespace OpenLoco::EntityManager
         }
     }
 
-    // 0x004A8826
-    void updateVehicles()
-    {
-        if (Game::hasFlags(GameStateFlags::tileManagerLoaded) && !isEditorMode())
-        {
-            for (auto v : VehicleList())
-            {
-                v->updateVehicle();
-            }
-        }
-    }
-
     // 0x004402F4
     void updateMiscEntities()
     {
@@ -321,21 +302,6 @@ namespace OpenLoco::EntityManager
             {
                 misc->update();
             }
-        }
-    }
-
-    // 0x004B94CF
-    void updateDaily()
-    {
-        call(0x004B94CF);
-    }
-
-    // 0x004C3C54
-    void updateMonthly()
-    {
-        for (auto v : VehicleList())
-        {
-            v->updateMonthly();
         }
     }
